refactor: move consts.ts to 'src/consts'
